// src/Interfaces/UnsplashResponse.tsx
export interface UnsplashResponse {
    id: string
    created_at: string
    updated_at: string
    promoted_at: string
    width: number
    height: number
    color: string
    blur_hash: string
    description: string
    alt_description: string
    urls: Urls
    links: Links
    likes: number
    liked_by_user: boolean
    current_user_collections: any[]
    sponsorship: any
    topic_submissions: TopicSubmissions
    user: User
    exif: Exif
    location: Location
    meta: Meta
    public_domain: boolean
    tags: Tag[]
    tags_preview: TagsPreview[]
    views: number
    downloads: number
    topics: Topic[]
  }
  
  export interface Urls {
    raw: string
    full: string
    regular: string
    small: string
    thumb: string
    small_s3: string
  }
  
  export interface Links {
    self: string
    html: string
    download: string
    download_location: string
  }
  
  export interface TopicSubmissions {
    wallpapers: Wallpapers
  }
  
  export interface Wallpapers {
    status: string
    approved_on: string
  }
  
  export interface User {
    id: string
    updated_at: string
    username: string
    name: string
    first_name: string
    last_name: string
    twitter_username: string
    portfolio_url: string
    bio: string
    location: string
    links: Links2
    profile_image: ProfileImage
    instagram_username: string
    total_collections: number
    total_likes: number
    total_photos: number
    accepted_tos: boolean
    for_hire: boolean
    social: Social
  }
  
  export interface Links2 {
    self: string
    html: string
    photos: string
    likes: string
    portfolio: string
    following: string
    followers: string
  }
  
  export interface ProfileImage {
    small: string
    medium: string
    large: string
  }
  
  export interface Social {
    instagram_username: string
    portfolio_url: string
    twitter_username: string
    paypal_email: any
  }
  
  export interface Exif {
    make: string
    model: string
    name: string
    exposure_time: string
    aperture: string
    focal_length: string
    iso: number
  }
  
  export interface Location {
    name: string
    city: string
    country: string
    position: Position
  }
  
  export interface Position {
    latitude: number
    longitude: number
  }
  
  export interface Meta {
    index: boolean
  }
  
  export interface Tag {
    type: string
    title: string
    source?: Source
  }
  
  export interface Source {
    ancestry: Ancestry
    title: string
    subtitle: string
    description: string
    meta_title: string
    meta_description: string
    cover_photo: CoverPhoto
  }
// ...
  export interface TagsPreview {
    type: string
    title: string
    source?: Source2
  }
  
  export interface Source2 {
    ancestry: Ancestry2
    title: string
    subtitle: string
    description: string
    meta_title: string
    meta_description: string
    cover_photo: CoverPhoto2
  }
// ...
  export interface Topic {
    id: string
    title: string
    slug: string
    visibility: string
  }